requirements.txt
flask
python-dotenv
azure-storage-blob
azure-identity